My neck no longer hurts :)
This commit adds a bunch of features, including:

* Use the correct accelerometer value scale

Not all devices have the accelerometer raw values between 0 and 1 (e.g. PineTab).
The correct scale is accessible via sysfs.

* Revamped keyboard mode system

You can specify how your keyboard is attached so that rot8 will do the
appropriate thing. `--keyboard-mode` can be one of:

- `integrated`: The keyboard is an integral part of the device. We will
  disable the keyboard when the screen is not oriented upright.
- `detachable`: The keyboard is detachable. We will keep the screen
  oriented upright when the keyboard is attached.
- `none`: We won't touch your keyboard.

The old `-k`/`--disable-keyboard` flag has been removed.

* After-rotation hook

Specify a command to be run in shell after rotation with `--rotate-hook`. Useful
for forcing a restart of Polybar, for example.

* PineTab-specific flags

There are several PineTab-specific additions. If you are using a
PineTab, the following flags should be given:

--display DSI-1 --touchscreen 'pointer:Goodix Capacitive TouchScreen' --flip-y
--rotate-90 --keyboard-mode detachable --keyboard 'HAILUCK CO.,LTD USB KEYBOARD'

If you are setting the default rotation in Xorg.conf, use `TransformationMatrix`
instead of `CalibrationMatrix` for the touchscreen, otherwise the setting will
conflict with rot8.

Issue: efernau/rot8#15